requirements.txt
absl-py==1.4.0
amulet-core==1.9.29
amulet-leveldb==1.0.2
amulet-map-editor==0.10.42
amulet-nbt==2.1.4
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
appdirs==1.4.4
appnope==0.1.3
asttokens==2.2.1
astunparse==1.6.3
async-generator==1.10
attrs==25.4.0
audioread==3.0.0
backcall==0.2.0
cachetools==6.2.6
celltk2==0.4.3
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.1.0
click==8.3.1
colorcet==3.0.1
comm==0.1.3
contourpy==1.2.0
cryptography==46.0.5
cycler==0.12.1
debugpy==1.6.7
decorator==5.1.1
Deprecated==1.3.1
deprecation==2.1.0
distro==1.9.0
exceptiongroup==1.3.1
executing==1.2.0
fastapi==0.135.2
flatbuffers==1.12
fonttools==4.49.0
fsspec==2026.2.0
gast==0.4.0
google-ai-generativelanguage==0.6.15
google-api-core==2.30.0
google-api-python-client==2.191.0
google-auth==2.49.1
google-auth-httplib2==0.3.0
google-auth-oauthlib==0.4.6
google-genai==1.68.0
google-generativeai==0.8.6
google-pasta==0.2.0
googleapis-common-protos==1.72.0
grpcio==1.78.0
grpcio-status==1.71.2
h11==0.16.0
h2==4.3.0
h5py==3.8.0
hpack==4.1.0
httpcore==1.0.9
httplib2==0.31.2
httptools==0.7.1
httpx==0.28.1
hyperframe==6.1.0
idna==3.4
imageio==2.34.0
importlib_metadata==8.7.1
iniconfig==2.3.0
ipykernel==6.22.0
ipython==8.12.0
jax==0.4.8
jedi==0.18.2
Jinja2==3.1.6
joblib==1.2.0
jupyter_client==8.2.0
jupyter_core==5.3.0
keras==2.9.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.5
lazy_loader==0.3
libclang==16.0.0
librosa==0.10.0.post2
limits==5.8.0
llvmlite==0.39.1
lz4==4.4.4
mahotas==1.4.12
Markdown==3.4.3
markdown-it-py==4.0.0
MarkupSafe==2.1.2
matplotlib==3.8.3
matplotlib-inline==0.1.6
mdurl==0.1.2
minecraft-resource-pack==1.4.6
ml-dtypes==0.1.0
mmh3==5.2.1
msgpack==1.0.5
multidict==6.7.1
mutf8==1.0.6
mypy_extensions==1.1.0
nest-asyncio==1.5.6
networkx==3.2.1
numba==0.56.4
numpy==1.26.4
oauthlib==3.2.2
opt-einsum==3.3.0
outcome==1.3.0.post0
packaging==25.0
pandas==2.0.1
param==2.0.2
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
pillow==11.2.1
platformdirs==3.11.0
plotly==5.19.0
pluggy==1.6.0
pooch==1.6.0
portalocker==2.10.1
postgrest==2.28.3
prompt-toolkit==3.0.38
propcache==0.4.1
proto-plus==1.27.1
protobuf==5.29.6
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==11.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pyct==0.5.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.15.1
pyiceberg==0.11.1
PyJWT==2.12.1
PyMCTranslate==1.2.33
PyOpenGL==3.1.9
pyparsing==3.1.1
pyroaring==1.0.4
PySocks==1.7.1
pytest==9.0.2
pytest-html==4.2.0
pytest-metadata==3.1.1
python-dateutil==2.8.2
python-dotenv==1.2.2
pytz==2023.3
PyYAML==6.0.1
pyzmq==25.0.2
realtime==2.28.3
requests==2.28.2
requests-oauthlib==1.3.1
resampy==0.4.2
rich==14.3.3
robotframework==7.4.1
robotframework-pythonlibcore==4.5.0
robotframework-seleniumlibrary==6.8.0
rsa==4.9
scikit-image==0.22.0
scikit-learn==1.2.2
scipy==1.10.1
seaborn==0.11.2
selenium==4.40.0
SimpleITK==2.3.1
six==1.16.0
slowapi==0.1.9
sniffio==1.3.1
sortedcontainers==2.4.0
soundfile==0.12.1
soxr==0.3.5
stack-data==0.6.2
starlette==1.0.0
storage3==2.28.3
StrEnum==0.4.15
strictyaml==1.7.3
supabase==2.28.3
supabase-auth==2.28.3
supabase-functions==2.28.3
tenacity==8.2.3
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-estimator==2.9.0
tensorflow-macos==2.9.2
tensorflow-metal==0.8.0
termcolor==2.2.0
tf-slim==1.1.0
threadpoolctl==3.1.0
tifffile==2024.2.12
tomli==2.4.0
tornado==6.3
tqdm==4.67.3
traitlets==5.9.0
trio==0.32.0
trio-typing==0.10.0
trio-websocket==0.12.2
types-certifi==2021.10.8.3
types-urllib3==1.26.25.14
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2023.3
uritemplate==4.2.0
urllib3==1.26.20
uvicorn==0.42.0
uvloop==0.22.1
watchfiles==1.1.1
wcwidth==0.2.6
websocket-client==1.9.0
websockets==15.0.1
Werkzeug==2.2.3
wrapt==1.14.1
wsproto==1.3.2
wxPython==4.2.3
yarl==1.23.0
zipp==3.23.0
zstandard==0.25.0
